requirements.txt
beautifulsoup4==4.11.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
cloudscraper==1.2.68
coverage==7.6.1
cycler==0.12.1
fonttools==4.54.1
idna==3.10
joblib==1.4.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lxml==4.9.1
matplotlib==3.6.2
mypy-extensions==1.0.0
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==24.1
pandas==1.5.1
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.3.6
pyparsing==3.1.4
pyTelegramBotAPI==4.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.28.1
requests-toolbelt==1.0.0
scikit-learn==1.1.3
scipy==1.7.1
six==1.16.0
soupsieve==2.6
threadpoolctl==3.5.0
tomli==2.0.2
torch==1.13.0
typing_extensions==4.12.2
urllib3==1.26.12
